+++ src/lib.rs
@@ -1,11 +1,45 @@
+//! output rust version info file
+//!
+//! This crate is the presents, the file output of rustc --version and cargo tree command.
+//!
+//! # Examples
+//!
+//! Please write the following code in the build.rs:
+//!
+//! ```rust
+//! use rust_version_info_file::rust_version_info_file;
+//! 
+//! fn main() {
+//!     rust_version_info_file("target/rust-version-info.txt", "Cargo.toml");
+//! }
+//! ```
+//!
+//! And you get the file as result it.
+//!
+//! ```text
+//! cat target/rust-version-info-file.txt
+//! ```
+//!
+//! # On debian package
+//! 
+//! In Cargo.toml
+//! 
+//! ```text
+//! [package.metadata.deb]
+//! assets = [
+//!     ["target/rust-version-info.txt", "usr/share/doc/your_package/", "644"],
+//!     ["README.md", "usr/share/doc/your_package/", "644"],
+//! ]
+//! ```
+//! 
 use std::io::{Read, Write};
 use std::path::Path;
 use std::fs::OpenOptions;
 
-pub fn rust_version_info_file<T: AsRef<Path>>(dst: T) {
+pub fn rust_version_info_file<T: AsRef<Path>>(dst: T, cargo_toml_file: &str) {
     let dst_path: &Path = dst.as_ref();
     let old_s = read_file(dst_path);
-    let curr_s = format!("{}\n{}\n", rustc_version_info(), tree_version_info(),);
+    let curr_s = format!("{}\n{}\n", rustc_version_info(), tree_version_info(cargo_toml_file),);
     if old_s != curr_s {
         let mut fo = match OpenOptions::new()
             .create(true)
@@ -51,10 +85,12 @@ fn rustc_version_info() -> String {
     String::from_utf8(v.to_vec()).unwrap()
 }
 
-fn tree_version_info() -> String {
+fn tree_version_info(cargo_toml_file: &str) -> String {
     let cmd = "cargo";
     let out = std::process::Command::new(cmd)
         .arg("tree")
+        .arg("--manifest-path")
+        .arg(cargo_toml_file)
         .output()
         .unwrap();
     let v = out.stdout;
@@ -64,5 +100,22 @@ fn tree_version_info() -> String {
     } else {
         &v[0..v_len]
     };
-    String::from_utf8(v.to_vec()).unwrap()
+    //
+    let string = String::from_utf8(v.to_vec()).unwrap();
+    //
+    let string = match string.find(" (") {
+        Some(pos1) => {
+            let (s1, s2) = string.split_at(pos1);
+            match s2.find(')') {
+                Some(pos2) => {
+                    let (_s2, s3) = s2.split_at(pos2+1);
+                    s1.to_owned() + s3
+                }
+                None => string,
+            }
+        },
+        None => string,
+    };
+    //
+    string
 }